Fix audit log viewer: keep server-only code out of the client bundle
Clicking Activity in the admin console died with 'Error loading route
module': the activity page's client code imports audit-query, which
imported AUDIT_ACTIONS from audit.ts, which imports rate-limit.ts — whose
module-level setInterval(...).unref() throws in a browser (unref is
Node-only). The compiled unref() call shipped in the production client
bundle, so the route was broken in deployed builds too, not just dev.

AUDIT_ACTIONS moves to a browser-safe audit-actions.ts (audit.ts
re-exports it for server callers), and rate-limit's unref becomes an
optional call so an accidental client-bundling of this module can never
crash a page at init again.
